push inline class to is own file
